modelos creados hasta Beneficiario

+++ src/main/java/com/bonoperubackend/BonoPeruBackend/Repositorios/ValoresRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ValoresRepository extends JpaRepository<Valores,Long> {
+public interface ValoresRepository extends JpaRepository<Valores,Integer> {
 
 }
